Разбила проект на модули
