--- src/view/film-list-button.js
@@ -1,2 +1,0 @@
-export const createFilmListButtonTemplate = () => '<button class="films-list__show-more">Show more</button>';
-
