+++ src/view/film-show-more-button.js
@@ -1,0 +1,2 @@
+const BUTTON_LABEL = 'Show more';
+export const createFilmListButtonTemplate = () => `<button class="films-list__show-more">${BUTTON_LABEL}</button>`;
